% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllPlots.R
\name{AllPlots}
\alias{AllPlots}
\title{Create a PDF file with lots of pot in a single page, to have a good  overwiew  of the run
Author: Thibaud Derippe}
\usage{
AllPlots(object, pdf = T, pdfall = F, covPrim = "none",
  covSec = "none", runDataDes = 0, addFunction = "",
  sensitivity = "")
}
\description{
Arguments: pdf if you want a pdf output, always true for the moment (I will probably delate this attribute)
Arguments: pdfall, the idea was to either have all plots in the same page, either one plot per page. Might be deleted
Note : pdf and pdfall are very old attribute and residues of old thoughts (i changed my mind concerning format extraction).
The code of extraction is thus over paramtrized and messy (but work as I would like to). I will simplified it
}
\details{
Arguments: covPrim the most important covariable you want your plot taking into account.
Arguments: covSec, if a second covariable has to be taken into account (mostly for boxplots)
Note : I will implement Non-Standard Evaluation in the future for covPrim and covSec

Arguments: addFunction to add a second page with personnal function. Probably a messy code as it's a recent adjunction
and I never checked the code since (I will do some adjustement in the future)

Output: a pdf with lots of plots
}
