% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2generics.R
\name{bining}
\alias{bining}
\title{bining
To transform a string of numbers into numbers (only used in Shiny app)}
\usage{
bining(vect, times1_group_2 = 1, step_ngroup)
}
\value{
la fonction renvoie la moyenne d'un vecteur
}
\description{
bining
To transform a string of numbers into numbers (only used in Shiny app)
}
\author{
Thibaud Derippe
}
