% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2generics.R
\name{compteur}
\alias{compteur}
\title{compteur
Function for auto-incrementation.  used for saving ggplots in list, before a do.call}
\usage{
compteur()
}
\value{
la fonction renvoie la moyenne d'un vecteur
}
\description{
compteur
Function for auto-incrementation.  used for saving ggplots in list, before a do.call
}
\author{
Thibaud Derippe
}
