% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3createRun_UserFunction.R
\name{createRun}
\alias{createRun}
\title{Create a run object with the pathway
Author: Thibaud Derippe
Arguments: path of the run
Names: don't remenber, I thing it's an old residu that can be removed?
Output: a run object}
\usage{
createRun(path)
}
\description{
Create a run object with the pathway
Author: Thibaud Derippe
Arguments: path of the run
Names: don't remenber, I thing it's an old residu that can be removed?
Output: a run object
}
