% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecc_ode.R
\docType{data}
\name{depot}
\alias{depot}
\title{ODE monolix}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.}
\usage{
depot
}
\arguments{
\item{func}{same as the ode}

\item{parms}{same as the ode}

\item{y}{same as the ode}

\item{datafile}{Either a .tab or an other cfl}

\item{...}{}
}
\description{
ODE monolix
}
\keyword{datasets}
