% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3SetReplace_Dossier.R
\docType{methods}
\name{dossier_fill_COMMUN<-,folder-method}
\alias{dossier_fill_COMMUN<-,folder-method}
\title{Fill the dossier with NONMEM or Monolix runs
Author: Thibaud Derippe
Arguments: object, the dossier object to use the setReplaceMethod
Arguments: value, the pathway to = begin the researches
Argumets: auto_storage if you want to auto-load runs(useless I think, can be deleted)
Output: not appropriate, setReplaceMethod}
\usage{
\S4method{dossier_fill_COMMUN}{folder}(object) <- value
}
\description{
Fill the dossier with NONMEM or Monolix runs
Author: Thibaud Derippe
Arguments: object, the dossier object to use the setReplaceMethod
Arguments: value, the pathway to = begin the researches
Argumets: auto_storage if you want to auto-load runs(useless I think, can be deleted)
Output: not appropriate, setReplaceMethod
}
