% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3find_run_while_loop.R
\name{find_runs}
\alias{find_runs}
\title{Giving a path of a folder and the software, it search every single runs in the folder
Author: Thibaud Derippe
Arguments: path, the pathway to begin the researches of runs
Output: a tibble containing the name of the runs and the complete pathway}
\usage{
find_runs(path)
}
\description{
Giving a path of a folder and the software, it search every single runs in the folder
Author: Thibaud Derippe
Arguments: path, the pathway to begin the researches of runs
Output: a tibble containing the name of the runs and the complete pathway
}
