% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DossierObject.R
\docType{class}
\name{folder-class}
\alias{folder-class}
\title{SetClass S4 Dossier Object
Author: Thibaud Derippe (\email{Thibaud.Derippe@gmail.com})}
\description{
SetClass S4 Dossier Object
Author: Thibaud Derippe (\email{Thibaud.Derippe@gmail.com})
}
\section{Slots}{

\describe{
\item{\code{RunsStorage}}{list of runs included in the dossier object}

\item{\code{Summary}}{summary of every runs the dossier will handle}

\item{\code{racine}}{root of the folder}

\item{\code{software}}{wether it's NONMEM or Monolix}

\item{\code{lastSelected}}{see how CreateDossier function works}
}}

