% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2generics.R
\name{ggplot_pecc}
\alias{ggplot_pecc}
\title{Ggplot + theme_bw + title/subtitle centered + caption left and italic}
\usage{
ggplot_pecc(data = NULL, mapping = aes())
}
\description{
Ggplot + theme_bw + title/subtitle centered + caption left and italic
}
