% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peccary_NCA.R
\name{NCA_covariate}
\alias{NCA_covariate}
\title{Non Compartimental Analysis  (Warning, ongoing development)}
\usage{
NCA_covariate(NCA_analysis, cov = "", DT_output = T, y_tag = "",
  x_tag = "", nsignif = 3, title = "")
}
\description{
Non Compartimental Analysis  (Warning, ongoing development)
}
\author{
Thibaud Derippe (\email{Thibaud.Derippe@gmail.com})
}
