% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peccary_NCA.R
\name{NCA_report}
\alias{NCA_report}
\title{Non Compartimental Analysis  (Warning, ongoing development)}
\usage{
NCA_report(dataset, x = TIME, y = IPRED, group = ID, occ = OCC,
  auc_0_X = 0, covariates = c("BATCH", "SEX"))
}
\description{
Non Compartimental Analysis  (Warning, ongoing development)
}
\author{
Thibaud Derippe (\email{Thibaud.Derippe@gmail.com})
}
