% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecc_count.R
\name{pecc_count}
\alias{pecc_count}
\title{Peccary count}
\usage{
pecc_count(df, metric, col1, col2)
}
\arguments{
\item{df}{Name of the dataframe to analyse}

\item{metric}{Name of the column metric (NSA)}

\item{col1}{Name of the first column}

\item{col2}{Name of the second column}
}
\description{
Take a dataframe, the name of two columns and a boolen one, compute percentage for each combination.
}
\examples{
pecc_count(mtcars,metric = vs, col1 = cyl, col2 = carb)
}
\author{
Thibaud Derippe
}
