% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecc_dataset_creation.R
\name{pecc_dataset_creation}
\alias{pecc_dataset_creation}
\title{Dataset creation for Monolix and NONMEM}
\usage{
pecc_dataset_creation(df, admin = NULL, time = NULL, obs = "C_ng_mL",
  dose = NULL, dose_fct = 1000, evid2 = NULL, LOQ = NULL)
}
\description{
Dataset creation for Monolix and NONMEM
}
