% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peccary_app.R
\name{peccary_app}
\alias{peccary_app}
\title{#' @export
dossier <- new("folder")
#' @export
explo <- tibble()
path <- ""
#' @export
plots_output <- tibble(Plot = character(), Runs = character(), plots = list(), Commentary = character())
launch peccary app}
\usage{
peccary_app()
}
\description{
#' @export
dossier <- new("folder")
#' @export
explo <- tibble()
path <- ""
#' @export
plots_output <- tibble(Plot = character(), Runs = character(), plots = list(), Commentary = character())
launch peccary app
}
