% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peccary_FO_cov.R
\docType{methods}
\name{peccary_FO_COV,folder-method}
\alias{peccary_FO_COV,folder-method}
\title{Objective Function Comparison through covariate}
\usage{
\S4method{peccary_FO_COV}{folder}(object, cov = "")
}
\description{
Objective Function Comparison through covariate
}
