% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peccary_NCA.R
\name{peccary_NCA}
\alias{peccary_NCA}
\title{Non Compartimental Analysis  (Warning, ongoing development)}
\usage{
peccary_NCA(dataset, x = TIME, y = IPRED, group = ID, occ = OCC,
  auc_0_X = 0, dose = dose, extend = F, npoint = 3,
  forceLinear = F, Tinf = 0)
}
\arguments{
\item{dataset}{data.frame/tibble containing all informations}

\item{x}{name of the x column  (TIME per default)}

\item{y}{name of the y column (IPRED per default)}

\item{group}{name of the group column (ID per default)}

\item{dose}{name of the dose column (dose per default)}

\item{extend}{if false perform a minimal analysis}

\item{npoint}{number of last point usef for last slope computation}

\item{forceLinear}{bool, if false (default) use a mixed Linear log-linear trapezoidal method, if false use only linear method}

\item{Tinf}{numeric, time of infusion, used for MRT correction}
}
\description{
compute NCA analysis from a dataset. Function on development, use with precaution.
}
\author{
Thibaud Derippe (\email{Thibaud.Derippe@gmail.com})
}
