% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_GOF.R
\name{plot_GOF}
\alias{plot_GOF}
\title{Goodness of Fits Plots}
\usage{
plot_GOF(object, plots_to_keep = F, cov_to_use = "", col_cov = F,
  gather_by_cov = F, plotly = F, filter = F, table_output = F,
  xlog = "", ylog = "", units_obs = "", units_time = "",
  method_smooth = "loess")
}
\arguments{
\item{object}{either a run or a folder}

\item{plots_to_keep}{by default  create every possible GOF. User can select only plots of interest either by numbers or by strings}

\item{cov_to_use}{string of the covariate header the user want to split GOFS}

\item{col_cov}{bool, if true make one color per covariate on a single plot, if F make one layer per covariate value}

\item{gather_by_cov}{bool, wether the user want plots gathered by covariate or by type of GOFS}

\item{plotly}{bool, if true transforms static plots into plotly dynamic plots, usefull to detect outliers}

\item{filter}{a filter can contains IPRED/PRED/OBS/TIME/NPDE/YTYPE... and any covariable (included ID) that you want. Can be quoted or not.}

\item{table_output}{TRUE or FALSE, either to directly print the plots or to return the table}
}
\description{
Create goodness of fits plots for one or several runs, eventually spread by a covariate.
}
\details{
plot_GOF() search of every possible goodness of fit plots and create them individually. If a proporitional error is found, according plots will
be in logarithmic scales. User can filter as it wants the dataset. If several runs is given in input, through a folder, it will give back the table of every plots doable. User can then
filter and arrange as it wants and then use the plot_invoke() function
}
\author{
Thibaud Derippe \email{thibaud.derippe@gmail.com}
}
