% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peccary_NCA.R
\name{plot_NCA}
\alias{plot_NCA}
\title{Non Compartimental Analysis  (Warning, ongoing development)}
\usage{
plot_NCA(NCA_analysis, n = 6)
}
\description{
Non Compartimental Analysis  (Warning, ongoing development)
}
\author{
Thibaud Derippe (\email{Thibaud.Derippe@gmail.com})
}
