% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Pred.R
\name{plot_pred}
\alias{plot_pred}
\title{Prediction Plots}
\usage{
plot_pred(object, filterr = "", covs = F, n = 20, pred = T,
  ylog = T, xlog = F, freeScale = F, plotly = F, LOQ = "")
}
\arguments{
\item{object}{peccary object, namely a run or a folder, generally through a folder function (with one or several run numbers).}

\item{filterr}{string or NSE, directly through a R code (for instance \emph{"IPRED > 3 & dose \%in\%  c(100, 200)"}, with or without quotes). Works with IPRED, PRED, OBS, TIME, YTYPE and any covariate (included ID).
This filter is secondly reworked so too complex filters can lead to errors - notably filters like (A & B) | (C & D)...}

\item{covs}{string, name of the covariate the user want to discriminate trough different colors. Works only for a single run (method not implemented for folders).}

\item{n}{numeric, number of maximal plots represented. If the number of possible ID is above this value, a sampling process is applied.}

\item{pred}{bool, display (or not) prediction profiles in dashed lines.}

\item{ylog}{bool, if true (by default) a logarithmic y axis is applied, if false an ordinary one.}

\item{xlog}{bool, if true a logarithmic x axis is applied, if false (by default) an ordinary one.}

\item{plotly}{bool, if true toggle a dynamic plot, if false (by default) display a static plot.}

\item{LOQ}{numeric, manually provide a LLOQ value. Every point below or equal to this value will be displayed with a different shape. A horitonzal dashed line will also be printed.}

\item{freescale}{bool, if true every ID will have its personnal axis, if false shared axis will be used.}
}
\value{
Returns a ggplot object (which can be reworked, see example)
}
\description{
Display individual and population predictions (IPRED and PRED) profiles versus observation dots over time for a given number of subjects (sampled or selected through filters).
}
\examples{
\dontrun{

# folder creation
folder <- createFolder("path_to_initial_folder")

# Basic plot_pred
plot_pred(folder(1), covs = "Dose", n = 25, LOQ = 10 )

# Applying a filter for a console peccary use:
plot_pred(folder(3:4),  IPRED > 0.1 & x_cov == "y")

# Equivalent to (for shiny app purpose mostly):
plot_pred(folder(3:4),  "IPRED > 0.1 & x_cov == \\"y\\"")

# Created plot can be manually modified:

 plot_pred(folder(1)) +
  ggtitle("Peccary is a beautifull animal") +
  theme_dark() +
  labs(x = "region", y = "peccary fertility")

}
}
\author{
Thibaud Derippe (\email{Thibaud.Derippe@gmail.com})
}
