% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_pred_dist.R
\name{plot_pred_dist}
\alias{plot_pred_dist}
\title{Plot Pred Dist}
\usage{
plot_pred_dist(dataset, sim, x, y, obs = "OBS10", group = "",
  ylog = T, loq_obs = 0.01, ymin_displayed = 1e-04, xlab = "Time",
  ylab = "Concentration", filterr = "", ...)
}
\arguments{
\item{dataset}{Name of the dataset containing simulations}

\item{sim}{Name of the simulation column (SIM, SIMUL...)}

\item{x}{Name of the x column (Time in general)}

\item{y}{Name of the y column (DV, IPRED....)}

\item{obs}{Name of the observation column (to be displayed)}

\item{group}{Name of one or several group to split (e.g. "AGE + SEX")}

\item{ylog}{Booléen, for logarithmique scale}

\item{loq_obs}{Value of loq for observation}

\item{loq_simul}{Value of loq for simulations (on test, use with precaution)}
}
\description{
Plot Pred Dist
}
\author{
Thibaud Derippe
}
