% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_random_effect_cov.R
\docType{methods}
\name{plot_random_effect_cov,run-method}
\alias{plot_random_effect_cov,run-method}
\title{BoxPlot according covariabe
Author: Thibaud Derippe
Arguments: object, a run or severals through a dossier
Arguments: covPrim : wich cov to use in x axis
Arguments: CovSec: ggplot color separation
ARguments: output : grid always (removed?)
Arguments: boxplotPoint if you want to add the observation point
Output: IndPred plots}
\usage{
\S4method{plot_random_effect_cov}{run}(object, covPrim = "none",
  covSec = "none", output = "grid", boxplotPoint = F,
  boxplotest = "anova")
}
\description{
BoxPlot according covariabe
Author: Thibaud Derippe
Arguments: object, a run or severals through a dossier
Arguments: covPrim : wich cov to use in x axis
Arguments: CovSec: ggplot color separation
ARguments: output : grid always (removed?)
Arguments: boxplotPoint if you want to add the observation point
Output: IndPred plots
}
