% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_random_effect_matrix.R
\docType{methods}
\name{plot_random_effect_matrix,run-method}
\alias{plot_random_effect_matrix,run-method}
\title{Random Effect Matrix
Author: Thibaud Derippe
Arguments: object, a run
Output: random effect matrix plots}
\usage{
\S4method{plot_random_effect_matrix}{run}(object, table = F,
  filterr = "")
}
\description{
Random Effect Matrix
Author: Thibaud Derippe
Arguments: object, a run
Output: random effect matrix plots
}
