% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Spagh.R
\name{plot_spagh}
\alias{plot_spagh}
\title{Spaghetti plot}
\usage{
plot_spagh(df, x = "TIME", y = "DV", id = "NMID2", col = "",
  facetwrap = "", facetgrid = "", facetscale = "free", median = T,
  point = 1, ind_alpha = 0.3, standardisation = "", ylog = T,
  xlog = F, loq = "", filter = "", title = "", caption = "")
}
\description{
Spaghetti plot
}
