% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Spagh.R
\name{plot_spagh2}
\alias{plot_spagh2}
\title{filter<- "YTYPE == 1"
facetgrid <- ""
facetwrap <- ""
standardisation <- ""
col <- "DOSE"
loq <- 2
#
plot_spagh2(df, loq = 0.001, x = TIME, y = DV, id = NMID2, col = DOSE, filter = YTYPE == 1, median = F, addlines = addlines)
# eval(parse_expr(plot_spagh2(df, x = "TIME", y = "DV", id = "NMID2", filter = "YTYPE == 1", median = F)))
# microbenchmark(plot_spagh2(df, x = "TIME", y = "DV", id = "NMID2", filter = "YTYPE == 1", median = F), plot_spagh(df, x = "TIME", y = "DV", id = "NMID2", filter = "YTYPE == 1", median = F))
#'Spaghetti plot
#'@export}
\usage{
plot_spagh2(df, x = "TIME", y = "DV", id = "NMID2", col = "",
  facetwrap = "", facetgrid = "", facetscale = "free", median = T,
  point = 1, ind_alpha = 0.3, standardisation = "DOSE_bckp",
  ylog = T, xlog = F, loq = "", filter = "", title = "",
  subtitle = "", caption = "", xlabs = "", ylabs = "",
  output = "plot", sizetext = 15, removeNA = T, addlines)
}
\description{
addlines <- tibble(type = factor(c("hline", "vline")), intercept = c("0:4", "c(1,50)"), lty = factor(c("dashed", "dashed")))
}
\details{
expr(tibble(!!!map(addlines %>% mutate_if(is.factor, as.character), ~.x)))
expr(tibble(!!!map(addlines, ~parse_expr(.x))))

addlines
# plot_spagh2(df2,facetwrap = "DOSE", x = "TIME", y = DV, id = "NMID2", col = "NMID2", filter = YTYPE == 1, median = F, output = "plot")
Spaghetti plot2
}
