% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_vpc.R
\name{plot_vpc}
\alias{plot_vpc}
\title{Visual Prediction Check}
\usage{
plot_vpc(dataset, x = "TIME", y = "IPRED", sim = "SIMUL",
  obs = "OBS", type_scatter_CI = "CI", quantiles = c(5, 50, 95),
  CIpct = 5, statobs = T, group = "", ylog = T, stat_obs = F,
  loq_obs = 0, xlabel = "", ylabel = "", ymin_displayed = 0,
  filterr = "")
}
\arguments{
\item{dataset}{Dataframe containing simulations}

\item{x}{Name of x columns (Time columns most of time)}

\item{y}{Name of Y columns (IPRED, DV...)}

\item{sim}{Name of simulation columns (SIM, SIMUL...)}

\item{obs}{Name of observation columns (to add dots)}

\item{group}{Name of one or several group to split (e.g. "AGE + SEX")}

\item{ylog}{Booléen, if the plot must be turned in logarithmic scale}

\item{stat_obs}{Booléen, if stats of observation (q5, q50, q95) must be displayed}

\item{loq_obs}{Value of loq for observation}

\item{xlabel}{Name of the label for x}

\item{ylabel}{Name of the label for y}

\item{ymin_displayed}{Value of loq for simulations (on test, use with precaution)}

\item{id}{Name of the column with ID}
}
\value{
A ggplot object (modifiable)
}
\description{
Compute visual predicted check plot
}
\author{
Thibaud Derippe
}
