% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2generics.R
\name{readTablePerso}
\alias{readTablePerso}
\title{readTablePerso}
\usage{
readTablePerso(path, header = T, skip = 0)
}
\arguments{
\item{numbers}{a string}
}
\description{
Because of monolix while using inter occasion... If read table doesn't work (beause of "#") I do a readLines then I transform it in table (which take times... probably better options exist)
}
\author{
Thibaud Derippe
}
