% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results.R
\docType{methods}
\name{results,run-method}
\alias{results,run-method}
\title{Create a df with every run selected}
\usage{
\S4method{results}{run}(object, plot = F)
}
\arguments{
\item{run}{}
}
\description{
Create a df with every run selected
}
