% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2runObject.R
\docType{class}
\name{run-class}
\alias{run-class}
\title{SetClass S4 run Object
Author: Thibaud Derippe}
\description{
SetClass S4 run Object
Author: Thibaud Derippe
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the run (usefull for plots title for instance)}

\item{\code{path}}{root of the run, such as with name + path we can have the complete pathway of the run}

\item{\code{status}}{usefull only for NONMEM: is the run successfully converged or did it failed (and a key work of the nature of the error)}

\item{\code{FO}}{: objective function value}

\item{\code{Estimation}}{: parameter estimation values with RSE}

\item{\code{OBS}}{: every observation}

\item{\code{IPRED}}{: every estimation}

\item{\code{cov}}{: ID every covariable concerning each ID}

\item{\code{randomEffect}}{: random effect values}

\item{\code{residus}}{: residual values}

\item{\code{administration}}{: sampling schedule}

\item{\code{loq}}{: loq values}

\item{\code{software}}{: NONMEM or Monolix}

\item{\code{date}}{: date of creation}
}}

