% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2generics.R
\name{runIndivi}
\alias{runIndivi}
\title{runIndivi
To transform a string of numbers into numbers (only used in Shiny app)}
\usage{
runIndivi(numbers = "1 2 5:10 - 6:7 9")
}
\arguments{
\item{numbers}{a string}
}
\value{
la fonction renvoie la moyenne d'un vecteur
}
\description{
runIndivi
To transform a string of numbers into numbers (only used in Shiny app)
}
\author{
Thibaud Derippe
}
