% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2generics.R
\name{scale_y_log10_pecc}
\alias{scale_y_log10_pecc}
\title{Ggplot + theme_bw + plot.(sub)title(element.text = hjust(0.5))}
\usage{
scale_y_log10_pecc()
}
\description{
Ggplot + theme_bw + plot.(sub)title(element.text = hjust(0.5))
}
