% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3SelectRunFromFolder.R
\docType{methods}
\name{select_run,folder-method}
\alias{select_run,folder-method}
\title{From a folder give the run
Author: Thibaud Derippe
Arguments: object, the dossier object
Arguments: value, the run you want to create from
Output: a run object}
\usage{
\S4method{select_run}{folder}(object, value = 0)
}
\description{
From a folder give the run
Author: Thibaud Derippe
Arguments: object, the dossier object
Arguments: value, the run you want to create from
Output: a run object
}
