% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3SetReplace_MONOLIX2016_Run.R
\name{setPath_MONOLIX<-}
\alias{setPath_MONOLIX<-}
\title{setPath_MONOLIX
Author: Thibaud Derippe}
\usage{
setPath_MONOLIX(object) <- value
}
\description{
setPath_MONOLIX
Author: Thibaud Derippe
}
