% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2generics.R
\name{stringToNumbers}
\alias{stringToNumbers}
\title{string to number
To transform a string of numbers into numbers (only used in Shiny app)}
\usage{
stringToNumbers(numbers)
}
\arguments{
\item{numbers}{a string}
}
\value{
la fonction renvoie la moyenne d'un vecteur
}
\description{
string to number
To transform a string of numbers into numbers (only used in Shiny app)
}
\author{
Thibaud Derippe
}
